feat(SpokePoolClient): Initial new event support (#865)
These new events are supported in test, but not in production. A
contract bump w/ new ABIs is still required for production support, but 
this commit at least frontloads a lot of the necessary changes.

nb. most of the SpokePoolClient changes here were pulled out of the 
existing branch bz/eventIngestion.