feat: spoke pool client queries old and new event names (#829)
This refactors the spoke pool client so that it may query events for the
upgraded spoke pool. We must be able to query and manage both events
since the spoke pool upgrade will occur asynchronously.

I believe the goal of this PR is to simply be able to index the new
events into our spoke pool clients. It is then up to the function/client
using the spoke pool clients to handle these new event structures.

We need to handle:

ExecutedRelayerRefundRoot - No change. Addresses/bytes32 are stored as
strings in RelayerRefundLeaf (see here)
EmergencyDeleteRootBundle - No change. We do not query this event in the
hub pool client.
TokensBridged - No change. The l2TokenAddress is kept as a string by us.
For the next four signatures, both events are queried using the same
logic as before, and the event data is assigned to a single state
variable. For example, when we query both V3FundsDeposited and
FundsDeposited, we create the internal DepositWithBlock struct out of
the event data and store all these events into this.depositHashes (so
this.depositHashes will have deposits with addresses as inputToken and
outputToken if it was processed from a V3FundsDeposited event, and will
also have deposits with inputToken and outputToken as bytes32 values if
it was processed from a FundsDeposited event.

V3FundsDeposited -> FundsDeposited
RequestedSpeedUpV3Deposit -> RequestedSpeedUpDeposit
FilledV3Relay -> FilledRelay
RequestedV3SlowFill -> RequestedSlowFill

Signed-off-by: bennett <bennett@umaproject.org>
Signed-off-by: Matt Rice <matthewcrice32@gmail.com>
Co-authored-by: Dong-Ha Kim <dongha.kim210@gmail.com>
Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>
Co-authored-by: Matt Rice <matthewcrice32@gmail.com>