fix(test): Correct testing of FilledRelay messageHash
This change fixes two issues:
- MockSpokePoolClient incorrectly bundles the deposit message field into
  the FilledRelay event.
- The SpokePoolClient incorrectly overwrites the FilledRelay event
  messageHash on update.

The combination of these meant that tests to detect incorrect
messageHash fields were ineffective because the FilledRelay messageHash
was incidentally overwritten with the exact intended value. Since the
former bug is not present in production RPCs, the latter bug is exposed
and messageHash fields are unconditionally set to bytes32(0).