feat: Add class to handle address retrieval
This commit introduces a class that can be used to retrieve a set of
addresses from one or more remote sources. Different backends can be
implemented to support different response formats.

Effort is made to be flexible with the responses received. Addresses are
not required to be checksummed; as long as they can be coerced into a
checksummed address then they will be accepted. Invalid addresses are
dropped without side-effects.