feat(multicall): Add blockAndAggregate (#915)
Multicall3.blockAndAggregate() is a nice utility that returns the block
number with the response. This can be useful when the blockTag to query
might be defined, which in many cases otherwise requires an initial
eth_getBlockByNumber("latest") call. blockAndAggregate should permit a
single RPC call to be saved in these instances.