feat(utils): Add findDepositBlock (#912)
This implements the bare minimum functionality needed to identify the 
block number for a deposit. It performs a similar function to
getBlockRangeForDepositId() but has no dependency on a SpokePoolClient
instance, and is less strict about the inputs. The underlying algorithm
is simple: find the lowest blockNumber where numberOfDeposits is greater
than the target deposit Id. This effectively mirrors the behaviour of
findFillBlock().

The main motivator for this is to push chain-specific knowledge out of
the SpokePoolClient in advance of non-EVM code being introduced.