fix(findDepositBlock): Drop blockAndAggregate
This unfortunately doesn't work as intended on Arbitrum One & Orbit
chains. These chains unfortunately return the approximate block number
of the _underlying_ chain. This is totally jarring and can only be fixed
with a custom Multicall3 adaptation for Arbitrum chains.

Maybe one day we'll use blockAndAggregate with a custom Multicall3
deployment on Arbitrum chains. For now, revert to something that works
as expected.