refactor(svm): Factor out individual event decoding (#968)
Isolate the dependency on Anchor to src/svm/utils/events.ts. Do this in
part by using ethers base58 and base64 implementations.

This was motivated by the need to add some utilities for fetching and
decoding SVM transactions within the relayer repository, but without
wanting to depend directly on anchor.