improve(SpokeUtils): Use viem for RelayData hash computation
The Viem ABI encoder permits supplying the ABI as an object, which is
faster than ethers, which apparently requires that the ABI be supplied
to AbiCoder.encode() as a string, which implies re-parsing the same ABI
on each call.

Brief testing showed that the new implementation is 50% faster than the
existing one. Producing 500 RelayData hashes with each approach took 0.5
seconds with viem and 1 second with ethers. This is relevant in the
relayer because performs bulk fillStatus queries of up to 250 RelayData
hashes per chain on update.