improve(TokenUtils): Replace getL1TokenInfo
This function is kind of dangerous in that it intuitively returns the `TokenInfo` for the L1 token matching the passed in l2 token/chain but it actually returns the L2 token decimals. I could see this function being misused in the future.

This PR will force users to use `getL1TokenAddress` to get the L1 token address for an l2 token and use `getTokenInfo` to get the decimals/symbol for the same l2 token.