chore(SpokePoolClient): Tweaks for downstream implementation
The "enhanced" listener classes down in the relayer need to deal with
extending both the EVM and SVM SpokePoolClient variants. This is working
well with Typescript mixins, but there are a couple of hurdles.

Primarily, the depositHashes member is protected in the
SpokePoolClient, but protected members cannot be declared in the mixin
type specification because they are inherently internal. Make
depositHashes globally readable for the time being.

Additionally, the SVMSpokePoolClient create static method cannot be used
with mixins because it narrows the returned type to an
SVMSPokePoolClient. Therefore, remove the protected constraint on the
constructor and instead just recommend users to use the create() method.

Ultimately, the SpokePoolClient listener functionality will need to
migrate upstream to the SDK, where it can be more directly integrated.
This will permit these workarounds to be dropped.