feat(NetworkUtils): Add chainIsOFTEnabled() (#1199)
Initial variant of a utility function to determine whether a token can
be bridged via OFT. The logic here is incomplete because OFT support is
a function of the chain and the token being bridged. This will need to
be refined in follow-up work.

While here, nuke a redundant chainIsBSC() function - this isn't used
anywhere and isn't required.