chore: Make contract check EIP-7702 aware (#1074)
isContractDeployedToAddress() currently mistakes EIP-7702 delegations
for contracts. For most/all purposes in Across, 7702 delegation is
preferred to be treated as an EOA. This change permits the caller to
specify whether EIP-7702 delegations should be considered as contract
code. EIP-7702 delegation is considered as a contract by default, so
this is backwards compatible with the existing implementation and must
requires opt-in.