refactor(solana): Improve RPC error detection
RPC errors must also be evaluated at the caching/aggregator layer to
avoid rotating to another provider in the event that the first provider
fails. Some failures cannot be resolved and need to be aborted
immediately.

While testing this, it was identified that SolanaError objects can be
mangled somewhere in the call stack, such that isSolanaError() doesn't
resolve true properly. It's unclear why this is the case, but introduce
a secondary option for resolving isSolanaError() via superstruct.