refactor(solana): Improve RPC error detection (#1251)
RPC errors must also be evaluated at the caching/aggregator layer to
minimise unnecessary rotations to another provider in the event that the 
first provider fails. Some failures cannot be resolved and need to be
aborted immediately.

Additionaly, the `isSolanaError` predicate can fail outside of the 
Solana RPC provider implementations. The underlying issue seems to be
that isSolanaError relies on instanceof, and some SolanaError properties
are getting slightly mangled as the errors are rippled up through the
various layers. This was traced to tryWithFallback() inside
QuorumFallbackSolanaRpcFactory, but I wasn't able to prevent it from
occurring. To mitigate this, implement a fallback discrimination that
infers SolanaError based on the shape of the object that's thrown.